--- a/extensions/vertx-http/runtime/src/main/java/io/quarkus/vertx/http/runtime/VertxInputStream.java
+++ b/extensions/vertx-http/runtime/src/main/java/io/quarkus/vertx/http/runtime/VertxInputStream.java
@@ -246,7 +246,7 @@ public class VertxInputStream extends InputStream {
             if (readException != null) {
                 throw readException;
             }
-            if (!endOfWrite && (inputOverflow.isEmpty())) {
+            if (!endOfWrite && inputOverflow.isEmpty()) {
                 request.fetch(1);
             }
             if (ret == null && !inputOverflow.isEmpty()) {