--- a/independent-projects/resteasy-reactive/client/runtime/src/main/java/org/jboss/resteasy/reactive/client/handlers/VertxClientInputStream.java
+++ b/independent-projects/resteasy-reactive/client/runtime/src/main/java/org/jboss/resteasy/reactive/client/handlers/VertxClientInputStream.java
@@ -222,7 +222,7 @@ class VertxClientInputStream extends InputStream {
             if (readException != null) {
                 throw readException;
             }
-            if (!endOfWrite && (inputOverflow.isEmpty())) {
+            if (!endOfWrite && inputOverflow.isEmpty()) {
                 request.fetch(1);
             }
             if (ret == null && !inputOverflow.isEmpty()) {