@fredericBregier 2023-11-25
Note 2: POST is about 4 times slower than GET, but still better than nothing. 
Note 3: Using Netty as low level client side, performances are 4 times better on GET, 50% better on POST. But that's another story.