@fredericBregier 2023-12-12
A bit improvement by fetching 3 instead of 1 at startup (so little memory consumption) and fetching each time a buffer is consumed (this seems to enable better pipelining between consumption and network):

POST: from 950 MB/s to 1200 MB/s, compare to Netty as client only up to 1400 MB/s
GET: from 950 MB/s to 2000 MB/s, compare to  Netty as client only up to 4400 MB/s

Note that increasing the startup fetching from 2 to 5 still increases the result, with still no memory impact (not noticeable):
- POST : 1) 950 MB/s 2) 1100 MB/s  3) 1200 MB/s 4) 1250 MB/s 5) 1300 MB/s
- GET: 1) 950 MB/s 2) 1500 MB/s 3) 2000 MB/s 4) 2250 MB/s 5) 2300 MB/s

My conviction is that 3 shall be the conservative option to get the best average performance while minimizing the memory impact (even if not noticeable). 4 might be the best choice however. (maybe an option for that ?)