@geoand 2023-12-14
@vietj can you have a look at this as it very intense on the Vert.x side?

Thanks