@fredericBregier 2023-12-21
@vietj This implementation is inspired from various projects (not mines).
Of course, I cannot be certain 100%. I tried several ways.  I have an issue sometimes, never in simple POST (one client -> one server), but when chaining (client -> (server -> internal client) -> another server). I tried to find out why, but didn't get the reason yet.

However, if there is another way, I can try if you have any idea.

The issue right now is that sending an inputstream (POST) from a client leads to all bytes in memory, which leads to OOME. On the reverse way, getting InputStream (GET) is working (modulo some optimizations that I proposed too, but can be split in 2 MR).