@fredericBregier 2023-12-21
@vietj And indeed, only in double POST (client to server to another server), it is indeed VertxBlockingInput that seems to block. But as it is under the wood, it seems difficult to bypass it.
(Note again: in single POST, no chaining, no issue)

Any idea ?