@fredericBregier 2023-12-21
Just to give you more details: 

I update the scenario to show the issue using a self hack to include a deployed version of this.

```
@RegisterRestClient
@Path(ROOT_NORMAL)
public interface ApiInterface  {

  @POST
  Uni<Response> simplePost(InputStream inputStream);
}

@RegisterRestClient
@Path(ROOT_SECOND)
public interface ApiSecondInterface  {

  @POST
  Uni<Response> simplePost(InputStream inputStream);
}

@ApplicationScoped
public class ApiClient {
  @RestClient
  ApiInterface apiInterface;

  public long sendInputStream(InputStream inputStream) {
    Response response = apiInterface.simplePost(inputStream).await().indefinitely();
    return Long.parseLong(response.getHeaderString(HEADER_MAX_SIZE));
  }
}

@ApplicationScoped
public class ApiSecondClient {
  @RestClient
  ApiSecondInterface apiInterface;

  public long sendInputStream(InputStream inputStream) {
    Response response = apiInterface.simplePost(inputStream).await().indefinitely();
    return Long.parseLong(response.getHeaderString(HEADER_MAX_SIZE));
  }
}

@ApplicationScoped
@Path(ROOT_NORMAL)
public class ApiServer implements ApiInterface {
  @Inject
  ApiSecondClient client;
  
  @Override
  @Blocking
  public Uni<Response> simplePost(final InputStream inputStream) {
    return Uni.createFrom().emitter(em -> {
      final Semaphore semaphore = new Semaphore(0);
      final AtomicLong atomicLong = new AtomicLong();
      new Thread() { // To not block IO Thread... but maybe a better way ? Maybe the reason ?
        public void run() {
          try {
            long max = client.sendInput(inputStream);
            atomicLong.set(max);
          } finally {
            semaphore.release();
          }
        }
      }.start();
      try {
        semaphore.acquire();
      } catch (InterruptedException e) {
        throw new RuntimeException(e);
      }
      em.complete(Response.ok().header(HEADER_MAX_SIZE, atomicLong.get()).build());
    });
  }
}

@ApplicationScoped
@Path(ROOT_SECOND)
public class ApiSecondServer implements ApiSecondInterface {
  @Override
  @Blocking
  public Uni<Response> simplePost(final InputStream inputStream) {
    return Uni.createFrom().emitter(em -> {
      long max = ReadAll.readAll(inputStream); // simple method that read everything and returns the total number of bytes
      em.complete(Response.ok().header(HEADER_MAX_SIZE, max).build());
    });
  }
}

@QuarkusTest
class ApiTest {
  private static final Logger LOGGER = Logger.getLogger(ApiTest.class);
  @Inject
  ApiClient apiClient;
  @Inject
  ApiSecondClient apiSecondClient;

  @Test
  void testSendingProxy() {
    InputStream inputStream = new FakeInputStream(SIZE_1GB); // See code where FakeInputStream is written
    long size = apiClient.sendInputStream(inputStream);
    LOGGER.infof("Length %d", size);
  }

  @Test
  void testSendingDirect() {
    InputStream inputStream = new FakeInputStream(SIZE_1GB);
    long size = apiSecondClient.sendInputStream(inputStream);
    LOGGER.infof("Length:  %d", size);
  }
}
```
Test testSendingProxy is running but a subset of the inputStream only (around 1/2 MB), while testSendingDirect is OK.

So before doing anything else, it shall be ok at least for that too, right?

I don't have any idea yet.